requirements.txt
asgiref==3.2.10
certifi==2020.11.8
chardet==3.0.4
Django==3.1.2
idna==2.10
loguru==0.5.3
marshmallow==3.9.1
petl==1.6.8
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
